TypeScript conversion of demo-card
